daq-2.2.2

+++ daq.spec
@@ -1,65 +1,56 @@
-%define version 2.0.6
+%define version 2.2.2
 
 Summary: Data Acquisition Library
 License: GNU General Public License
 Group: Libraries/Network
 Name: daq
-Prefix: %{_prefix}
 Provides: daq
-Release: 1
+Release: 1%{dist}
 Source: daq-%{version}.tar.gz
 URL: http://www.snort.org/
 Version: %{version}
 
-BuildRoot: /tmp/daqrpm-%{version}
-BuildRequires: autoconf, automake, flex, libpcap-devel
+BuildRoot: %{_tmppath}/%{name}-%{version}-%{release}-root-%(%{__id_u} -n)
+Requires: libpcap
+BuildRequires: autoconf, automake, flex, gcc, bison, libpcap-devel
 
 %description
 Data Acquisition library for Snort.
+
+%package devel
+Summary: Development tools for programs to Data Acquisition (DAQ) Library
+
+%description devel
+Development tools for programs to Data Acquisition (DAQ) Library
+
 
 %prep
 %setup -q
 
 %build
 %configure
-make
+%{__make}
 
 %install
-rm -rf $RPM_BUILD_ROOT
-make install DESTDIR=$RPM_BUILD_ROOT
+%{make_install}
 
 %clean
-rm -rf $RPM_BUILD_ROOT
+[ -n "$RPM_BUILD_ROOT" -a "$RPM_BUILD_ROOT" != / ] && rm -rf $RPM_BUILD_ROOT
 
 %files
 %defattr(-,root,root)
-%{_includedir}/daq.h
-%{_includedir}/daq_common.h
-%{_includedir}/sfbpf_dlt.h
-%{_includedir}/daq_api.h
-%{_includedir}/sfbpf.h
-%{_libdir}/libdaq_static_modules.la
-%{_libdir}/libdaq_static_modules.a
-%{_libdir}/libsfbpf.a
-%{_libdir}/libdaq.so
-%{_libdir}/libsfbpf.la
-%{_libdir}/libdaq.la
-%{_libdir}/daq
-%{_libdir}/daq/daq_dump.so
-%{_libdir}/daq/daq_ipfw.la
-%{_libdir}/daq/daq_dump.la
-%{_libdir}/daq/daq_afpacket.la
-%{_libdir}/daq/daq_pcap.so
-%{_libdir}/daq/daq_pcap.la
-%{_libdir}/daq/daq_ipfw.so
-%{_libdir}/daq/daq_afpacket.so
-%{_libdir}/libsfbpf.so.0.0.1
-%{_libdir}/libsfbpf.so.0
-%{_libdir}/libdaq_static.la
-%{_libdir}/libdaq.so.2.0.4
-%{_libdir}/libsfbpf.so
-%{_libdir}/libdaq.a
-%{_libdir}/libdaq_static.a
-%{_libdir}/libdaq.so.2
-%{_bindir}/daq-modules-config
+%{_libdir}/*.so*
+%{_libdir}/daq/*.so*
+%doc ChangeLog
+%doc COPYING
+%doc README
 
+%files devel
+%attr(0755,root,root) %{_bindir}/daq-modules-config
+%{_includedir}/*.h
+%{_libdir}/*.{a,la}
+%{_libdir}/daq/*.{a,la}
+
+%changelog
+* Sun Mar 3 2019 Nobuyuki Ito <info@2ito.com> - 2.2.2-1
+- Fixed spec file for 2.2.2 and Fedora 29
